% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adj_sigmas}
\alias{adj_sigmas}
\title{Adjust Covariance Matrizes}
\usage{
adj_sigmas(adjust_sigmas = 1, fit, n, l, SIGMA_list, trace = 1)
}
\arguments{
\item{adjust_sigmas}{Should variances and/or covariances be adjusted? Default is 1 = adjust only variances, 2 = adjust both, NULL = no adjustments.}

\item{fit}{LDA fit with covariance functions}

\item{n}{number of cases}

\item{l}{index of variable for which covariance matrix should be adjusted}

\item{SIGMA_list}{List containing covariance matrizes for L}

\item{trace}{Controls console output. NULL = no output, 1 = basic output,
2 = detailed output.}
}
\description{
Helper function that adjusts only variances or both variances and covariances.
}
