% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_SIGMA_cond}
\alias{calc_SIGMA_cond}
\title{Conditional Covariance Matrix}
\usage{
calc_SIGMA_cond(ind_new, n_all_j, L, SIGMA, SIGMA_nn)
}
\arguments{
\item{ind_new}{Index in combined block diagonal Matrix to identify new, i.e.
test cases}

\item{n_all_j}{Number of all included cases: test cases to be predicted,
training cases included in conditional prediction}

\item{L}{Number of latent variables}

\item{SIGMA}{Combined block diagonal covariance Matrix}

\item{SIGMA_nn}{Subset of SIGMA, contains only values for test cases to be
predicted}
}
\description{
Computes a covariance matrix conditional on another subset of
the same multivariate normal distribution.
}
