% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalLDA-package.R
\docType{package}
\name{conditionalLDA-package}
\alias{conditionalLDA-package}
\title{Spatial LDA and conditional prediction}
\description{
This package implements spatial LDA model fit including covariance function
modelling. Predictions can include nearby training observations by means of a
conditional multivariate normal distribution (conditional prediction).
Predictions can be made for a whole block instead of single observations if
block identifiers are given.
}
