% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{f_sigmas}
\alias{f_sigmas}
\title{Get Covariances2}
\usage{
f_sigmas(x, covs, j)
}
\arguments{
\item{x}{index}

\item{covs}{Object to retrieve covariances from}

\item{j}{Index of block/case to be predicted}
}
\description{
Retrieves covariances for a certain variable
}
