% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_sp_helpers.R
\name{get_best_covmodel}
\alias{get_best_covmodel}
\title{Best covariance model}
\usage{
get_best_covmodel(
  coords,
  l,
  rs,
  omega0,
  ini = NULL,
  max_dist,
  trace,
  cov_models,
  showplot = TRUE,
  phi_max = NULL
)
}
\arguments{
\item{coords}{Dataframe with two rows containing coordinates}

\item{l}{The index of the used latent variables}

\item{rs}{List containing the residuals for each latent variable}

\item{omega0}{Covariance matrizes for all latent variables}

\item{max_dist}{Maximum distance between all points}

\item{trace}{Controls console output. NULL = no output, 1 = basic output,
2 = detailed output.}

\item{cov_models}{Covariance function models which are to be compared.}
}
\description{
Fits different given covariance function to the empirical semivariogram and chooses the best regarding sum of squares.
}
