% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_helpers.R
\name{get_nearby_obs}
\alias{get_nearby_obs}
\title{Get nearby observations}
\usage{
get_nearby_obs(fit, L, index, n_new_j, Y_new, coords_old, coords_new, cut_off)
}
\arguments{
\item{fit}{LDA fit providing coords of training data.}

\item{L}{Number of latent variables.}

\item{index}{indicates the cases belonging to the block to be predicted.}

\item{n_new_j}{number of cases belonging to the block to be predicted.}

\item{Y_new}{Latent variables/predictors of test data.}

\item{coords_old}{Coordinates of training data.}

\item{coords_new}{Coordinates of test set/cases to be classified.}

\item{cut_off}{Distance threshold up to which training observations are included.}
}
\description{
Identify training observations that lie within a given distance
treshold.
}
