% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_sp.R
\name{lda_sp}
\alias{lda_sp}
\title{Spatial LDA}
\usage{
lda_sp(
  formula,
  data,
  covs_equal = TRUE,
  coords = c("x", "y"),
  trace = 1,
  cov_models = c("matern", "exponential", "gaussian", "spherical"),
  max_dist = NULL,
  phi_max = NULL,
  showplot = TRUE,
  ini = NULL,
  fit = NULL
)
}
\arguments{
\item{formula}{of the form \emph{class} ~ x1 + x2 + ... that describes the response,
i.e. the class membership and the predictors on the right hand side.}

\item{data}{data frame containing the variables in the formula.}

\item{covs_equal}{Do the classes have equal covariance matrices? (LDA assumption)
OR Should class covariances be treated as equal?}

\item{coords}{vector of length 2 defining the variables in data that contain
x and y coordinates of sample locations. The default is c("x", "y"). Must be
given to perform spatial LDA.}

\item{trace}{Controls console output. NULL = no output, 1 = basic output,
2 = detailed output.}

\item{cov_models}{Character vector defining covariance function models which
are to be compared for each latent variable. Default is c("matern", "exponential",
"gaussian", "spherical"). Any combination of those as well as "linear",
"circular", "cubic", "wave" and ??"power"?? is possible. If set to NULL non
spatial/regular LDA will be performed.}
}
\description{
Fits a spatial LDA model that includes covariance
functions for each latent variable.
}
