% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliks.R
\name{loglik}
\alias{loglik}
\title{Loglikelihood}
\usage{
loglik(y, mu, C, logdet = NULL)
}
\arguments{
\item{y}{dimension reduced variable values}

\item{mu}{class means}

\item{C}{Covariance Matrix}

\item{logdet}{logarithm of matrix determinant}
}
\description{
Computes loglikelihood for given parameters of a multivariate normal distribution.
}
