% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliks.R
\name{loglik_train_sp}
\alias{loglik_train_sp}
\title{Loglikelihood training}
\usage{
loglik_train_sp(fit, L, adjust_sigmas)
}
\arguments{
\item{fit}{LDA fit with covariance functions and coords of training data}

\item{L}{Number of latent variables}

\item{adjust_sigmas}{Should variances and/or covariances be adjusted? Default is 1 = adjust both, 2 = adjust only variances, NULL = no adjustments.}
}
\description{
Computes loglikelihood of training observations
}
