% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cond.R
\name{predict_cond}
\alias{predict_cond}
\title{Conditional LDA Prediction}
\usage{
predict_cond(
  object,
  newdata,
  coords = c("x", "y"),
  adjust_sigmas = 1,
  trace = 1,
  LL = TRUE,
  blocks = "field",
  prior = TRUE,
  max_dist = NULL,
  corr_min = 0.01
)
}
\arguments{
\item{object}{LDA fit (object of class lda?) Has to be provide covariance functions and training data information for conditional prediction.}

\item{newdata}{data frame of cases to be classified.}

\item{coords}{vector of length 2 defining the variables in newdata that contain
x and y coordinates of sample locations. The default is c("x", "y"). Must be
given to perform conditional prediction.}

\item{adjust_sigmas}{Should variances and/or covariances be adjusted? Default is 1 = adjust only variances, 2 = adjust both, NULL = no adjustments.}

\item{trace}{Controls console output: 0 = no output, 1 = basic output , 2 = detailed output.}

\item{LL}{Should loglikelihood for training data be estimated?}

\item{blocks}{character string defining variable of newdata containing block identifiers.}

\item{prior}{If TRUE (default) predictions including prior class probabilities are returned.}
}
\value{
Predicted class memberships of given data.
}
\description{
This function predicts class membership for an entire block (or a single observation).
The conditional prediction approach includes nearby training observations.
Therefore coords must be given and LDA fit must provide covariance functions.
}
