% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_helpers.R
\name{predict_in}
\alias{predict_in}
\title{Non conditional Prediction}
\usage{
predict_in(fit, y_new, trace = FALSE, K)
}
\arguments{
\item{fit}{LDA fit}

\item{y_new}{latent variable values for all cases belonging to the block to be predicted}

\item{trace}{Controls console output. NULL = no output, 1 = basic output,
2 = detailed output.}
}
\description{
Helper function used within predict_cond for NON conditional predictions.
}
