% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_helpers.R
\name{predict_in_cond}
\alias{predict_in_cond}
\title{Conditional Prediction}
\usage{
predict_in_cond(
  fit,
  Y_new_j,
  Y_old_j,
  coords_new_j,
  coords_old_j,
  classes_old_j,
  trace,
  adjust_sigmas = 1
)
}
\arguments{
\item{fit}{LDA fit providing covariance functions of training data}

\item{Y_new_j}{latent variable values for all cases belonging to the block to be predicted}

\item{Y_old_j}{latent variable values for all training data cases within distance threshold}

\item{coords_new_j}{coords of all cases belonging to the block to be predicted}

\item{coords_old_j}{coord of all training data cases within distance threshold}

\item{classes_old_j}{class memberships of all training data cases within distance threshold}

\item{trace}{Controls console output. NULL = no output, 1 = basic output,
2 = detailed output.}

\item{adjust_sigmas}{Should variances and/or covariances be adjusted? Default is 1 = adjust only variances, 2 = adjust both, NULL = no adjustments.}
}
\description{
Helper function used within predict_cond for conditional predictions.
}
